+++ src/tenants/andalucia/components/LandingSection.tsx
