+++ src/tenants/andalucia/styles/anilloverdegranada.module.scss
@@ -118,7 +118,7 @@
 
 // Sections 2
 .section2MainContainer {
-    background-image: url('/tenants/andalusia/images/parallaxCover.jpg') ;
+    background-image: url('/tenants/andalucia/images/parallaxCover.jpg') ;
     width: 100vw;
     height: 240px;
     background-position: center;
